type cleanup
