+++ types/src/execution.rs
@@ -7,6 +7,7 @@ mod execution_result_v2;
 mod transform;
 mod transform_error;
 mod transform_kind;
+mod executor_query_request;
 
 pub use effects::Effects;
 pub use execution_result::ExecutionResult;
@@ -15,3 +16,4 @@ pub use execution_result_v2::ExecutionResultV2;
 pub use transform::TransformV2;
 pub use transform_error::TransformError;
 pub use transform_kind::{TransformInstruction, TransformKindV2};
+pub use executor_query_request::{ExecutorQueryRequestBuilder, ExecutorQueryRequest, ExecutorQueryResult};
